fix: completely debugged traversal tests
